requirements/dev.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --allow-unsafe --output-file=requirements/dev.txt --strip-extras requirements/raw/dev.txt
#
-e ./benchmarks
    # via -r requirements/raw/bench.txt
-e ./tests/tests_helpers
    # via -r requirements/raw/test_extra_none.txt
alabaster==0.7.13
    # via sphinx
annotated-types==0.6.0
    # via pydantic
astpath==0.9.1
    # via
    #   -r requirements/raw/lint.txt
    #   astpath
astroid==3.0.2
    # via pylint
attrs==23.2.0
    # via
    #   -r requirements/raw/test_extra_new.txt
    #   cattrs
babel==2.14.0
    # via sphinx
bandit==1.7.6
    # via -r requirements/raw/lint.txt
beautifulsoup4==4.12.2
    # via furo
build==1.0.3
    # via pip-tools
cachetools==5.3.2
    # via tox
cattrs==23.1.2
    # via -r requirements/raw/bench.txt
certifi==2023.11.17
    # via requests
cfgv==3.4.0
    # via pre-commit
chardet==5.2.0
    # via tox
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   pip-tools
    #   towncrier
colorama==0.4.6
    # via
    #   radon
    #   tox
contourpy==1.2.0
    # via matplotlib
coverage==7.3.3
    # via -r requirements/raw/dev.txt
cycler==0.12.1
    # via matplotlib
dataclass-factory==2.16
    # via -r requirements/raw/bench.txt
dill==0.3.7
    # via pylint
dirty-equals==0.7.1.post0
    # via -r requirements/raw/test_extra_none.txt
distlib==0.3.8
    # via virtualenv
dlint==0.14.1
    # via -r requirements/raw/lint.txt
docutils==0.20.1
    # via
    #   docutils-stubs
    #   myst-parser
    #   sphinx
    #   sphinx-better-subsection
    #   sphinx-paramlinks
docutils-stubs==0.0.22
    # via -r requirements/raw/doc.txt
filelock==3.13.1
    # via
    #   tox
    #   virtualenv
flake8==6.1.0
    # via dlint
fonttools==4.47.0
    # via matplotlib
furo==2023.9.10
    # via -r requirements/raw/doc.txt
gitdb==4.0.11
    # via gitpython
gitpython==3.1.40
    # via
    #   -r requirements/raw/doc.txt
    #   bandit
greenlet==3.0.2
    # via
    #   -r requirements/raw/test_extra_none.txt
    #   sqlalchemy
identify==2.5.33
    # via pre-commit
idna==3.6
    # via requests
imagesize==1.4.1
    # via sphinx
incremental==22.10.0
    # via towncrier
iniconfig==2.0.0
    # via pytest
invoke==2.2.0
    # via -r requirements/raw/runner.txt
isort==5.13.2
    # via pylint
jinja2==3.1.2
    # via
    #   myst-parser
    #   sphinx
    #   towncrier
kiwisolver==1.4.5
    # via matplotlib
lxml==5.0.0
    # via astpath
mando==0.7.1
    # via radon
markdown-it-py==3.0.0
    # via
    #   mdit-py-plugins
    #   myst-parser
    #   rich
markupsafe==2.1.3
    # via jinja2
marshmallow==3.20.1
    # via -r requirements/raw/bench.txt
mashumaro==3.10
    # via -r requirements/raw/bench.txt
matplotlib==3.8.2
    # via -r requirements/raw/_bench-orchestrate.txt
mccabe==0.7.0
    # via
    #   flake8
    #   pylint
mdit-py-plugins==0.4.0
    # via myst-parser
mdurl==0.1.2
    # via markdown-it-py
msgspec==0.18.4 ; implementation_name != "pypy"
    # via -r requirements/raw/bench.txt
mypy==1.7.1
    # via -r requirements/raw/lint.txt
mypy-extensions==1.0.0
    # via mypy
myst-parser==2.0.0
    # via -r requirements/raw/doc.txt
nodeenv==1.8.0
    # via pre-commit
numpy==1.26.3
    # via
    #   contourpy
    #   matplotlib
packaging==23.2
    # via
    #   build
    #   marshmallow
    #   matplotlib
    #   plotly
    #   pyproject-api
    #   pytest
    #   sphinx
    #   tox
pbr==6.0.0
    # via
    #   sphinxcontrib-apidoc
    #   stevedore
phonenumberslite==8.13.26
    # via -r requirements/raw/test_extra_none.txt
pillow==10.2.0
    # via matplotlib
pip-tools==7.3.0
    # via -r requirements/raw/dev.txt
platformdirs==4.1.0
    # via
    #   pylint
    #   tox
    #   virtualenv
plotly==5.18.0
    # via
    #   -r requirements/raw/_bench-orchestrate.txt
    #   -r requirements/raw/doc.txt
pluggy==1.3.0
    # via
    #   pytest
    #   tox
pre-commit==3.6.0
    # via -r requirements/raw/lint.txt
psutil==5.9.5
    # via
    #   -r requirements/raw/bench.txt
    #   pyperf
pycodestyle==2.11.1
    # via flake8
pydantic==2.4.2
    # via -r requirements/raw/bench.txt
pydantic-core==2.10.1
    # via pydantic
pyflakes==3.1.0
    # via flake8
pygments==2.17.2
    # via
    #   furo
    #   rich
    #   sphinx
pylint==3.0.3
    # via -r requirements/raw/lint.txt
pyparsing==3.1.1
    # via matplotlib
pyperf==2.6.1
    # via -r requirements/raw/bench.txt
pyproject-api==1.6.1
    # via tox
pyproject-hooks==1.0.0
    # via build
pytest==7.4.2
    # via
    #   -r requirements/raw/bench.txt
    #   -r requirements/raw/test_extra_none.txt
python-dateutil==2.8.2
    # via matplotlib
pytz==2023.3.post1
    # via dirty-equals
pyyaml==6.0.1
    # via
    #   bandit
    #   myst-parser
    #   pre-commit
radon==6.0.1
    # via -r requirements/raw/lint.txt
requests==2.31.0
    # via sphinx
rich==13.7.0
    # via bandit
ruff==0.3.0
    # via -r requirements/raw/lint.txt
schematics==2.1.1
    # via -r requirements/raw/bench.txt
six==1.16.0
    # via
    #   mando
    #   python-dateutil
smmap==5.0.1
    # via gitdb
snowballstemmer==2.2.0
    # via sphinx
soupsieve==2.5
    # via beautifulsoup4
sphinx==7.2.6
    # via
    #   -r requirements/raw/doc.txt
    #   furo
    #   myst-parser
    #   sphinx-basic-ng
    #   sphinx-copybutton
    #   sphinx-design
    #   sphinx-paramlinks
    #   sphinx-reredirects
    #   sphinxcontrib-apidoc
    #   sphinxcontrib-applehelp
    #   sphinxcontrib-devhelp
    #   sphinxcontrib-htmlhelp
    #   sphinxcontrib-qthelp
    #   sphinxcontrib-serializinghtml
    #   sphinxext-opengraph
sphinx-basic-ng==1.0.0b2
    # via furo
sphinx-better-subsection==0.2
    # via -r requirements/raw/doc.txt
sphinx-copybutton==0.5.2
    # via -r requirements/raw/doc.txt
sphinx-design==0.5.0
    # via -r requirements/raw/doc.txt
sphinx-paramlinks==0.6.0
    # via -r requirements/raw/doc.txt
sphinx-reredirects==0.1.3
    # via -r requirements/raw/doc.txt
sphinxcontrib-apidoc==0.4.0
    # via -r requirements/raw/doc.txt
sphinxcontrib-applehelp==1.0.7
    # via sphinx
sphinxcontrib-devhelp==1.0.5
    # via sphinx
sphinxcontrib-htmlhelp==2.0.4
    # via sphinx
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-qthelp==1.0.6
    # via sphinx
sphinxcontrib-serializinghtml==1.1.9
    # via sphinx
sphinxext-opengraph==0.9.1
    # via -r requirements/raw/doc.txt
sqlalchemy==2.0.27
    # via
    #   -r requirements/raw/test_extra_new.txt
    #   -r requirements/raw/test_extra_none.txt
stevedore==5.1.0
    # via bandit
tenacity==8.2.3
    # via plotly
toml==0.10.2
    # via vulture
tomlkit==0.12.3
    # via pylint
towncrier==23.11.0
    # via -r requirements/raw/dev.txt
tox==4.11.4
    # via -r requirements/raw/runner.txt
typing-extensions==4.9.0
    # via
    #   -r requirements/raw/test_extra_none.txt
    #   mashumaro
    #   mypy
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
urllib3==2.1.0
    # via requests
virtualenv==20.25.0
    # via
    #   pre-commit
    #   tox
vulture==2.10
    # via -r requirements/raw/lint.txt
wheel==0.42.0
    # via
    #   -r requirements/raw/pre.txt
    #   pip-tools

# The following packages are considered to be unsafe in a requirements file:
pip==23.3.1
    # via
    #   -r requirements/raw/pre.txt
    #   pip-tools
setuptools==69.0.0
    # via
    #   -r requirements/raw/dev.txt
    #   nodeenv
    #   pip-tools
